requirements.txt
fastapi
uvicorn
sqlalchemy
passlib
pydantic
psycopg2-binary
python-jose
python-multipart
bcrypt~=4.0.1
python-dotenv
supabase
gunicorn
requests